Initial Scaffolds

+++ db/migrate/20241110185206_create_products.rb
@@ -3,8 +3,8 @@ class CreateProducts < ActiveRecord::Migration[8.0]
     create_table :products do |t|
       t.string :name
       t.text :description
-      t.float :cost_price
-      t.float :retail_price
+      t.decimal :cost_price
+      t.decimal :retail_price
 
       t.timestamps
     end
